[AIRFLOW-3022] Fix unit test DAG `example_kubernetes_executor` is broken. Thus, put the unit test in `example_kubernetes_executor_config` (renamed from `example_kubernetes_annotation`)
