+++ tests/contrib/minikube/test_kubernetes_executor.py
@@ -158,55 +158,53 @@ class KubernetesExecutorTest(unittest.TestCase):
 
     def test_integration_run_dag(self):
         host = get_minikube_host()
-        dag_ids = ['example_kubernetes_annotation', 'example_kubernetes_executor']
+        dag_id = 'example_kubernetes_executor_config'
 
-        for dag_id in dag_ids:
-            result_json = self.start_dag(dag_id=dag_id, host=host)
+        result_json = self.start_dag(dag_id=dag_id, host=host)
 
-            self.assertGreater(len(result_json['items']), 0)
+        self.assertGreater(len(result_json['items']), 0)
 
-            execution_date = result_json['items'][0]['execution_date']
-            print("Found the job with execution date {}".format(execution_date))
+        execution_date = result_json['items'][0]['execution_date']
+        print("Found the job with execution date {}".format(execution_date))
 
-            # Wait 100 seconds for the operator to complete
-            self.monitor_task(host=host,
-                              execution_date=execution_date,
-                              dag_id=dag_id,
-                              task_id='start_task',
-                              expected_final_state='success', timeout=100)
+        # Wait 100 seconds for the operator to complete
+        self.monitor_task(host=host,
+                          execution_date=execution_date,
+                          dag_id=dag_id,
+                          task_id='start_task',
+                          expected_final_state='success', timeout=100)
 
-            self.ensure_dag_expected_state(host=host,
-                                           execution_date=execution_date,
-                                           dag_id=dag_id,
-                                           expected_final_state='success', timeout=100)
+        self.ensure_dag_expected_state(host=host,
+                                       execution_date=execution_date,
+                                       dag_id=dag_id,
+                                       expected_final_state='success', timeout=100)
 
     def test_integration_run_dag_with_scheduler_failure(self):
         host = get_minikube_host()
-        dag_ids = ['example_kubernetes_annotation', 'example_kubernetes_executor']
+        dag_id = 'example_kubernetes_executor_config'
 
-        for dag_id in dag_ids:
-            result_json = self.start_dag(dag_id=dag_id, host=host)
+        result_json = self.start_dag(dag_id=dag_id, host=host)
 
-            self.assertGreater(len(result_json['items']), 0)
+        self.assertGreater(len(result_json['items']), 0)
 
-            execution_date = result_json['items'][0]['execution_date']
-            print("Found the job with execution date {}".format(execution_date))
+        execution_date = result_json['items'][0]['execution_date']
+        print("Found the job with execution date {}".format(execution_date))
 
-            self._delete_airflow_pod()
+        self._delete_airflow_pod()
 
-            time.sleep(10)  # give time for pod to restart
+        time.sleep(10)  # give time for pod to restart
 
-            # Wait 100 seconds for the operator to complete
-            self.monitor_task(host=host,
-                              execution_date=execution_date,
-                              dag_id=dag_id,
-                              task_id='start_task',
-                              expected_final_state='success', timeout=120)
+        # Wait 100 seconds for the operator to complete
+        self.monitor_task(host=host,
+                          execution_date=execution_date,
+                          dag_id=dag_id,
+                          task_id='start_task',
+                          expected_final_state='success', timeout=120)
 
-            self.ensure_dag_expected_state(host=host,
-                                           execution_date=execution_date,
-                                           dag_id=dag_id,
-                                           expected_final_state='success', timeout=100)
+        self.ensure_dag_expected_state(host=host,
+                                       execution_date=execution_date,
+                                       dag_id=dag_id,
+                                       expected_final_state='success', timeout=100)
 
 
 if __name__ == '__main__':
